fix(1#): refactoring
