+++ src/steps/owner/constants.ts
@@ -1,0 +1,5 @@
+export const OwnerEntity = {
+  RESOURCE_NAME: 'Owner',
+  TYPE: 'acme_user',
+  CLASS: 'User',
+};
